Initial Recources Commit
This contains some placeholder texture and rights of publication stay with the original developers of the textures.

As progress continues original resources will replace those. The original Resources and Code that is created by me remain the intellectual property of Myself. if you wish to use them for personal learning you are welcome to. however, you can not redistribute them in any form or productionalise them without my express permission